Adds malakies

+++ controllers/cards.js
@@ -1,12 +1,13 @@
-const mongoose = require('mongoose');
 const Card = require('../models/card');
+const handleError = require('../middleware/handleError');
+const MestoError = require('../errors/MestoError');
 
 module.exports.getCards = (_, res) => {
   Card.find({})
     .populate('owner')
     .populate('likes')
     .then((cards) => res.send({ data: cards }))
-    .catch(() => res.status(500).send({ message: 'На сервере произошла ошибка' }));
+    .catch(() => handleError(res, new MestoError()));
 };
 
 module.exports.postCard = (req, res) => {
@@ -14,34 +15,21 @@ module.exports.postCard = (req, res) => {
   const { _id } = req.user;
   Card.create({ name, link, owner: _id })
     .then((card) => res.send({ data: card }))
-    .catch((err) => {
-      if (err instanceof mongoose.Error.ValidationError) {
-        res.status(400).send({ message: err.message });
-      } else {
-        res.status(500).send({ message: 'На сервере произошла ошибка' });
-      }
-    });
+    .catch((err) => handleError(res, err));
 };
 
 module.exports.deleteCard = (req, res) => {
-  Card.deleteOne({ _id: req.params.cardId })
+  Card.deleteOne({ _id: req.params.cardId, owner: req.user })
     .then((data) => {
       if (data.ok !== 1) {
         res.status(500).send({ message: 'На сервере произошла ошибка' });
-        return;
+        return Promise.reject(new MestoError());
       }
       if (data.deletedCount === 1) {
-        res.send({ message: 'Карта успешно удалена' });
-      } else {
-        res.status(404).send({ message: `Карта с id ${req.params.cardId} не найдена!` });
+        return res.send({ message: 'Карта успешно удалена' });
       }
-    }).catch((err) => {
-      if (err instanceof mongoose.Error.CastError) {
-        res.status(400).send({ message: `Передан некорректный идентификатор ${req.params.cardId}` });
-      } else {
-        res.status(500).send({ message: 'На сервере произошла ошибка' });
-      }
-    });
+      return Promise.reject(new MestoError(404, `вы не являетесь владельцем карты с идентификатором id ${req.params.cardId} !`));
+    }).catch((err) => handleError(res, err));
 };
 
 module.exports.likeCard = (req, res) => {
@@ -52,23 +40,12 @@ module.exports.likeCard = (req, res) => {
     { new: true },
   ).populate('owner').populate('likes')
     .then((card) => {
-      if (card) {
-        res.send({ data: card });
-      } else {
-        res.status(404).send({ message: `Карта с id ${req.params.cardId} не найдена!` });
+      if (!card) {
+        return Promise.reject(new MestoError(404, `Карта с id ${req.params.cardId} не найдена!`));
       }
+      return res.send({ data: card });
     })
-    .catch((err) => {
-      if (err instanceof mongoose.Error.CastError) {
-        if (err.path === '_id') {
-          res.status(400).send({ message: `Передан некорректный идентификатор ${req.params.cardId}` });
-        } else {
-          res.status(400).send({ message: `Передан некорректный идентификатор ${_id}` });
-        }
-      } else {
-        res.status(500).send({ message: 'На сервере произошла ошибка' });
-      }
-    });
+    .catch((err) => handleError(res, err));
 };
 
 module.exports.dislikeCard = (req, res) => {
@@ -79,21 +56,10 @@ module.exports.dislikeCard = (req, res) => {
     { new: true },
   ).populate('owner').populate('likes')
     .then((card) => {
-      if (card) {
-        res.send({ data: card });
-      } else {
-        res.status(404).send({ message: `Карта с id ${req.params.cardId} не найдена!` });
+      if (!card) {
+        return Promise.reject(new MestoError(404, `Карта с id ${req.params.cardId} не найдена!`));
       }
+      return res.send({ data: card });
     })
-    .catch((err) => {
-      if (err instanceof mongoose.Error.CastError) {
-        if (err.path === '_id') {
-          res.status(400).send({ message: `Передан некорректный идентификатор ${req.params.cardId}` });
-        } else {
-          res.status(400).send({ message: `Передан некорректный идентификатор ${_id}` });
-        }
-      } else {
-        res.status(500).send({ message: 'На сервере произошла ошибка' });
-      }
-    });
+    .catch((err) => handleError(res, err));
 };
